+++ araqne-logstorage/src/main/java/org/araqne/logstorage/file/LogFileReaderV2.java
@@ -179,7 +179,7 @@ public class LogFileReaderV2 extends LogFileReader {
 	}
 
 	@Override
-	public List<LogRecord> find(List<Integer> ids) throws IOException {
+	public List<LogRecord> find(List<Long> ids) throws IOException {
 		List<LogRecord> ret = new ArrayList<LogRecord>(ids.size());
 		
 		for (long id : ids) {
